+++ Sources/OEVoice/AVAudioSession+setSpeechSession.swift
@@ -9,11 +9,7 @@ import AVFoundation
 
 public extension AVAudioSession {
     /// Sets audioSession to play on mute, pause other spoken audio and duck anything else.
-    func setSpeechSession() {
-        do {
-            try self.setCategory(.playback, options: [.interruptSpokenAudioAndMixWithOthers, .duckOthers])
-        } catch {
-            print("Error: \(error.localizedDescription)")
-        }
+    func setSpeechSession() throws {
+        try self.setCategory(.playback, options: [.interruptSpokenAudioAndMixWithOthers, .duckOthers])
     }
 }